Added map functionality to the dashboard
